Move find-adslots
